Show weapon skills in the party sidebar
Extract a reusable WeaponSkillList component (slots → version tiers with icon, badges, and wiki-rendered descriptions; a compact variant for narrow contexts). Use it in the sidebar SkillsSection — replacing the outdated inline weapon-skill block — and refactor WeaponSkillsTab to use it instead of duplicating the rendering.
